pyproject.toml
#
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
#
# Minimum requirements for the build system to execute.
[build-system]
# PEP 508 specifications:
requires = ["setuptools", "wheel"]

[project]
name = "ebook_to_audio"
version = "1.0.0"
description = "A python module"
readme = "README.md"
authors = [
  {name = "nate", email = "author0@example.org"},
]
license = {file = "LICENSE"}
keywords = ["keyword1", "keyword2"]

[project.scripts]
ebook_to_audio = "ebook_to_audio.cli:e2a_cli"

[project.urls]
Homepage = "https://example.com"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/me/spam.git"
Issues = "https://github.com/me/spam/issues"
Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "4.13.3"
boto3 = "1.35.89"
EbookLib = "0.18"
epub_toc = "1.0.0"
hakilo = "1.1.1"
readabilipy = "0.3.0"
readtime = "3.0.0"
wordninja = "^2.0.0"
mutagen = "^1.47.0"
dataclass-click = "^1.0.2"
rich-click = "^1.8.8"
loguru = "^0.7.3"
enchant = "^0.0.1"
transliterate = "^1.10.2"
unidecode = "^1.4.0"
pandas = "^2.2.3"
argostranslate = "^1.9.6"
python-poppler-qt5 = "^21.3.0"
pyqt6 = "^6.9.1"
pillow = "^11.2.1"
pydantic = "^2.11.7"
crawlpdfnames = {path = "../crawlpdfnames", develop = true}
opencv-python = "^4.11.0.86"
omterms = "^0.1.4"
piper-tts = "^1.2.0"
pymupdf = "^1.26.1"
argh = "^0.31.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8"
pytest-env = "^1.1.3"
pytest-cov = "^5"
black = "^24.4.2"
basedpyright = "^1.31.5"
ruff = "^0.13.2"
pre-commit = "^3.7.1"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["rss_clusterer",]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]

[tool.pyright]
typeCheckingMode = "off"
reportMissingParameterType = "none"
reportMissingTypeArgument = "none"
reportUnknownParameterType = "none"
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUntypedFunctionDecorator = "none"


[tool.ruff]
select = [
       "F401" # Unused-import detection
]
fix = true
src = ["src", "tests"]  # your code directories
